Rename io.bitsquare.{util=>gui}.ViewLoader
