+++ src/main/java/io/bitsquare/gui/main/account/settings/AccountSettingsViewCB.java
@@ -21,9 +21,9 @@ import io.bitsquare.gui.CachedViewCB;
 import io.bitsquare.gui.Navigation;
 import io.bitsquare.gui.PresentationModel;
 import io.bitsquare.gui.ViewCB;
+import io.bitsquare.gui.ViewLoader;
 import io.bitsquare.gui.main.account.content.ContextAware;
 import io.bitsquare.gui.util.Colors;
-import io.bitsquare.util.ViewLoader;
 
 import java.net.URL;
