+++ src/test/java/io/bitsquare/gui/ViewLoaderTest.java
@@ -15,11 +15,10 @@
  * along with Bitsquare. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.bitsquare.util;
+package io.bitsquare.gui;
 
 import io.bitsquare.FatalException;
 import io.bitsquare.di.BitsquareModule;
-import io.bitsquare.gui.Navigation;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
